+++ src/components/__tests__/Loading-test.js
@@ -3,19 +3,171 @@ import {mount} from 'enzyme'
 import React from 'react'
 
 describe('Loading', () => {
-  it('functions as expected when type is not set', () => {
-    expect(mount(<Loading />)).toMatchSnapshot()
+  describe('when type is not set', () => {
+    let wrapper
+
+    beforeEach(() => {
+      wrapper = mount(<Loading />)
+    })
+
+    it('functions as expected', () => {
+      expect(wrapper).toMatchSnapshot()
+    })
+
+    describe('when type is set to null', () => {
+      beforeEach(() => {
+        wrapper.setProps({type: null})
+      })
+
+      it('functions as expected', () => {
+        expect(wrapper).toMatchSnapshot()
+      })
+    })
+
+    describe('when type is set to ring', () => {
+      beforeEach(() => {
+        wrapper.setProps({type: Loading.TYPES.RING})
+      })
+
+      it('functions as expected', () => {
+        expect(wrapper).toMatchSnapshot()
+      })
+    })
+
+    describe('when type is set to ripple', () => {
+      beforeEach(() => {
+        wrapper.setProps({type: Loading.TYPES.RIPPLE})
+      })
+
+      it('functions as expected', () => {
+        expect(wrapper).toMatchSnapshot()
+      })
+    })
   })
 
-  it('functions as expected when type is null', () => {
-    expect(mount(<Loading type={null} />)).toMatchSnapshot()
+  describe('when type is null', () => {
+    let wrapper
+
+    beforeEach(() => {
+      wrapper = mount(<Loading type={null} />)
+    })
+
+    it('functions as expected', () => {
+      expect(wrapper).toMatchSnapshot()
+    })
+
+    describe('when type is unset (set to undefined)', () => {
+      beforeEach(() => {
+        wrapper.setProps({type: undefined})
+      })
+
+      it('functions as expected', () => {
+        expect(wrapper).toMatchSnapshot()
+      })
+    })
+
+    describe('when type is set to ring', () => {
+      beforeEach(() => {
+        wrapper.setProps({type: Loading.TYPES.RING})
+      })
+
+      it('functions as expected', () => {
+        expect(wrapper).toMatchSnapshot()
+      })
+    })
+
+    describe('when type is set to ripple', () => {
+      beforeEach(() => {
+        wrapper.setProps({type: Loading.TYPES.RIPPLE})
+      })
+
+      it('functions as expected', () => {
+        expect(wrapper).toMatchSnapshot()
+      })
+    })
   })
 
-  it('functions as expected when type is ring', () => {
-    expect(mount(<Loading type={Loading.TYPES.RING} />)).toMatchSnapshot()
+  describe('when type is ring', () => {
+    let wrapper
+
+    beforeEach(() => {
+      wrapper = mount(<Loading type={Loading.TYPES.RING} />)
+    })
+
+    it('functions as expected', () => {
+      expect(wrapper).toMatchSnapshot()
+    })
+
+    describe('when type is unset (set to undefined)', () => {
+      beforeEach(() => {
+        wrapper.setProps({type: undefined})
+      })
+
+      it('functions as expected', () => {
+        expect(wrapper).toMatchSnapshot()
+      })
+    })
+
+    describe('when type is set to null', () => {
+      beforeEach(() => {
+        wrapper.setProps({type: null})
+      })
+
+      it('functions as expected', () => {
+        expect(wrapper).toMatchSnapshot()
+      })
+    })
+
+    describe('when type is set to ripple', () => {
+      beforeEach(() => {
+        wrapper.setProps({type: Loading.TYPES.RIPPLE})
+      })
+
+      it('functions as expected', () => {
+        expect(wrapper).toMatchSnapshot()
+      })
+    })
   })
 
-  it('functions as expected when type is ripple', () => {
-    expect(mount(<Loading type={Loading.TYPES.RIPPLE} />)).toMatchSnapshot()
+  describe('when type is ripple', () => {
+    let wrapper
+
+    beforeEach(() => {
+      wrapper = mount(<Loading type={Loading.TYPES.RIPPLE} />)
+    })
+
+    it('functions as expected', () => {
+      expect(wrapper).toMatchSnapshot()
+    })
+
+    describe('when type is unset (set to undefined)', () => {
+      beforeEach(() => {
+        wrapper.setProps({type: undefined})
+      })
+
+      it('functions as expected', () => {
+        expect(wrapper).toMatchSnapshot()
+      })
+    })
+
+    describe('when type is set to null', () => {
+      beforeEach(() => {
+        wrapper.setProps({type: null})
+      })
+
+      it('functions as expected', () => {
+        expect(wrapper).toMatchSnapshot()
+      })
+    })
+
+    describe('when type is set to ring', () => {
+      beforeEach(() => {
+        wrapper.setProps({type: Loading.TYPES.RING})
+      })
+
+      it('functions as expected', () => {
+        expect(wrapper).toMatchSnapshot()
+      })
+    })
   })
 })